update Product api implementation

+++ src/main/java/com/ranushka/cafe_management_system/wrapper/ProductWrapper.java
@@ -7,7 +7,7 @@ public class ProductWrapper {
     Integer id;
     String name;
     String description;
-    double price;
+    Integer price;
     String status;
     Integer categoryId;
     String categoryName;
@@ -15,7 +15,7 @@ public class ProductWrapper {
     public ProductWrapper() {
     }
 
-    public ProductWrapper(Integer id, String name, String description, double price, String status, Integer categoryId, String categoryName) {
+    public ProductWrapper(Integer id, String name, String description, Integer price, String status, Integer categoryId, String categoryName) {
         this.id = id;
         this.name = name;
         this.description = description;
